+++ src/react/components/autocompleter/XUIAutocompleterSecondarySearch.js
@@ -141,6 +141,7 @@ export default class XUIAutocompleterSecondarySearch extends PureComponent {
 
   render() {
     const {
+      _useCellStyling,
       qaHook,
       dropdownSize,
       dropdownClassName,
@@ -232,7 +233,11 @@ export default class XUIAutocompleterSecondarySearch extends PureComponent {
     );
 
     return (
-      <div className={className} data-automationid={containerQaHook} ref={this.rootNode}>
+      <div
+        className={cn(className, _useCellStyling && `${ns}-select-cell`)}
+        data-automationid={containerQaHook}
+        ref={this.rootNode}
+      >
         <XUIDropdownToggled
           className={dropdownToggledClasses}
           closeOnSelect={closeOnSelect}
@@ -253,6 +258,12 @@ export default class XUIAutocompleterSecondarySearch extends PureComponent {
 }
 
 XUIAutocompleterSecondarySearch.propTypes = {
+  /**
+   * @ignore
+   * Internal use only, used to assist with styling a button to look like part of a table
+   */
+  _useCellStyling: PropTypes.bool,
+
   children: PropTypes.node,
 
   /** CSS class(es) to go on the wrapping DOM node */
@@ -320,12 +331,14 @@ XUIAutocompleterSecondarySearch.propTypes = {
   // isLoading: PropTypes.bool,
 
   /**
-   * Setting to true will allow the dropdown's width to be set dependent of the trigger width. <br>
-   * **Note:** *Setting this to true will override any size prop on `XUIDropdown`.* <br>
+   * Setting this to `true` makes the dropdown as wide as the trigger. <br>
+   * **Note:** *Setting this to `true` will override any `size` prop on Dropdown.* <br>
+   * Setting this to `false` will allow the dropdown's width to be set independent of the trigger width (defaults to `false`). <br>
+   * Setting this to `'min'` will set the dropdown's `min-width` to be the trigger width. <br/>
    * XUI design has also decided to keep a minimum width on the dropdown,
-   * so dropdown may not match the width of narrow triggers.
+   * so dropdown may not match the width of narrow triggers (setting this to `'min'` will not override this).
    */
-  matchTriggerWidth: PropTypes.bool,
+  matchTriggerWidth: PropTypes.oneOfType([PropTypes.bool, PropTypes.oneOf([true, false, 'min'])]),
 
   /** Callback for when the list closes */
   onClose: PropTypes.func,